+++ CHANGELOG.md
@@ -9,6 +9,7 @@ NOT RELEASED
 #### Features
 * Content promotion through `Carousel` layout with editorial control.
 * `Featured` promotion through a topside, stretchable banner carousel.
+* Additional *"carousels"* added to menu
 
 #### Bug Fixes
 * Offline media list now reloads when the view appears